Decouple Rust code from generated data
