+++ site/src/quacs-rs/src/utils.rs
@@ -1,7 +1,5 @@
 #![allow(unused_macros)]
 
-use super::BIT_VEC_LEN;
-
 macro_rules! console_log {
     ( $( $t:tt )* ) => {
         if cfg!(feature = "debug") {
@@ -24,25 +22,25 @@ macro_rules! bm_end {
     };
 }
 
-pub fn bitwise_and(t1: &[u64; BIT_VEC_LEN], t2: &[u64; BIT_VEC_LEN]) -> [u64; BIT_VEC_LEN] {
-    let mut ret = [0; BIT_VEC_LEN];
-    for i in 0..BIT_VEC_LEN {
+pub fn bitwise_and<const N: usize>(t1: &[u64; N], t2: &[u64; N]) -> [u64; N] {
+    let mut ret = [0; N];
+    for i in 0..N {
         ret[i] = t1[i] & t2[i];
     }
     ret
 }
 
-pub fn bitwise_or(t1: &[u64; BIT_VEC_LEN], t2: &[u64; BIT_VEC_LEN]) -> [u64; BIT_VEC_LEN] {
-    let mut ret = [0; BIT_VEC_LEN];
-    for i in 0..BIT_VEC_LEN {
+pub fn bitwise_or<const N: usize>(t1: &[u64; N], t2: &[u64; N]) -> [u64; N] {
+    let mut ret = [0; N];
+    for i in 0..N {
         ret[i] = t1[i] | t2[i];
     }
     ret
 }
 
-pub fn bitwise_xor(t1: &[u64; BIT_VEC_LEN], t2: &[u64; BIT_VEC_LEN]) -> [u64; BIT_VEC_LEN] {
-    let mut ret = [0; BIT_VEC_LEN];
-    for i in 0..BIT_VEC_LEN {
+pub fn bitwise_xor<const N: usize>(t1: &[u64; N], t2: &[u64; N]) -> [u64; N] {
+    let mut ret = [0; N];
+    for i in 0..N {
         ret[i] = t1[i] ^ t2[i];
     }
     ret
